Ultimos ajustes nesta bagaça antes de botar pra f*der

+++ src/main/java/br/com/urcontroler/data/entity/Menu.java
@@ -42,7 +42,7 @@ public class Menu implements Comparable<Menu> {
      */
     public Menu(Long id, String title, String icon) {
         this.id = id;
-        this.parent = (long) 0;
+        this.parent = 0l;
         this.title = title;
         this.icon = icon;
     }
